Refactor recipe fetch call to Recipes controller
